Don't use 'with' statement for device DB connections
Previously, when creating a connection to the device DB, it was done
from 'with' statements and wrapped in 'closing()' calls. Since calling
close() on DB connections is unnecessary those closing() calls were
removed. However, this had the side effect of now calling the connection
itself as a context manager, creating a potentially unwanted
transaction. This caused VACUUM commands to fail, since they can't be
run from within a transaction.

This commit removes all 'with' statements from connection creations that
previously were wrapped in closing(). This restores the previous
behaviour of not creating a transaction and also saves an indentation
level.

Also remove all unnecessary cursor.close() calls.

Fixes: #5