Handle new device database infrastructure in Calibre 8
In Calibre 8 device databases may be copied to a temporary location due
to filesystem limitations, for example Chromebooks may mount the device
using the 9p filesystem. In order to prevent data loss if Calibre and
the plugin use different database locations, we need to get the copied
database location from the device driver and also lock the DB during
transactions.