Check for Qt typing issues
Add Qt as a dependency so that it can be properly typechecked.
Fix a whole lot of Qt issues as a result of this, mostly from changes in
Qt6. Calibre has compatibility code that still makes Qt5 things work in
practice, but it's better not to rely on this, and the typechecker
doesn't know about it.