Save config dialog device info in instance var
Previously, the device info was saved in the first column of each row in
the device table. This made it difficult to update and persist the info,
because you had to fiddle with the table implementation all the time.

Saving the devices as a copy of the plugin config at the start of the
dialog makes it easy to update things on the fly, and assign it back to
the plugin config when persisting after the user clicks "OK".